Wait to publish until AFTER the function config is updated
The way AWS uses $LATEST is very confusing. It is NOT a pointer to the
latest version (as you might think). Instead it is a copy of whatever version
you are publishing.

so it seems to me like:
- Whenever you upload code, it creates a version called $LATEST (in addition to whatever version you're publishing)
- When you update configuration, it only updates config for the $LATEST version
- When publish a new version, it uses the configuration from $LATEST
so if you publish a new version, then modify the $LATEST config, it doesn't modify your new version's config
You have to modify $LATEST, and _then_  publish a new version to get the config